Move tests from test_serialize_sequence.cc to test_SequenceManager.cc
[why]
The moved tests rely heavily (and sometimes exclusively) on the
SequenceManager class, so they should be located with the other
SequenceManager tests.
