+++ tests/test_serialize_sequence.cc
@@ -23,17 +23,12 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 #include <algorithm>
-#include <chrono>
 #include <filesystem>
 #include <sstream>
-#include <system_error>
-#include <vector>
 
 #include <gul14/catch.h>
 
-#include "../src/internals.h" // SEQUENCE_LUA_FILENAME
 #include "taskolib/deserialize_sequence.h"
-#include "taskolib/SequenceManager.h"
 #include "taskolib/serialize_sequence.h"
 
 using namespace std::literals;
@@ -43,29 +38,6 @@ using namespace task::literals;
 namespace {
 
 static const std::filesystem::path temp_dir{ "unit_test_files" };
-
-// Helper that returns all Lua step entries of a directory (i.e. files or subdirs)
-std::vector<std::string> collect_lua_filenames(const std::filesystem::path& path)
-{
-    std::vector<std::string> result;
-    for (const auto& entry: std::filesystem::directory_iterator{ path })
-        if (entry.path().extension() == ".lua")
-            result.push_back(entry.path().filename().string());
-
-    std::sort(result.begin(), result.end());
-    return result;
-}
-
-// Helper that returns all entries of a directory (i.e. files or subdirs)
-std::vector<std::string> collect_filenames(const std::filesystem::path& path)
-{
-    std::vector<std::string> result;
-    for (const auto& entry: std::filesystem::directory_iterator{ path })
-        result.push_back(entry.path().filename().string());
-
-    std::sort(result.begin(), result.end());
-    return result;
-}
 
 } // anonymous namespace
 
@@ -427,281 +399,4 @@ R"(
 
     Step deserialize;
     REQUIRE_THROWS_AS(ss >> deserialize, Error);
-}
-
-TEST_CASE("serialize_sequence: test filename format", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    const SequenceName seq_name{ "Test_sequence" };
-    const UniqueId seq_uid{ 0xdeadbeef };
-    const auto seq_folder = make_sequence_filename(seq_name, seq_uid);
-
-    if (std::filesystem::exists(temp_dir / seq_folder))
-        std::filesystem::remove_all(temp_dir / seq_folder);
-
-    REQUIRE_THROWS_AS(load_sequence(temp_dir / seq_folder), Error);
-
-    Step step01{Step::type_action};
-    step01.set_label("action");
-    Step step02{Step::type_if};
-    step02.set_label("if");
-    Step step03{Step::type_action};
-    step03.set_label("action");
-    Step step04{Step::type_elseif};
-    step04.set_label("elseif");
-    Step step05{Step::type_action};
-    step05.set_label("action");
-    Step step06{Step::type_end};
-    step06.set_label("end");
-    Step step07{Step::type_while};
-    step07.set_label("while");
-    Step step08{Step::type_action};
-    step08.set_label("action");
-    Step step09{Step::type_end};
-    step09.set_label("end");
-    Step step10{Step::type_action};
-    step10.set_label("action");
-
-    Sequence sequence{ "", seq_name, seq_uid };
-    sequence.push_back(step01);
-    sequence.push_back(step02);
-    sequence.push_back(step03);
-    sequence.push_back(step04);
-    sequence.push_back(step05);
-    sequence.push_back(step06);
-    sequence.push_back(step07);
-    sequence.push_back(step08);
-    sequence.push_back(step09);
-    sequence.push_back(step10);
-
-    REQUIRE_NOTHROW(manager.store_sequence(sequence));
-
-    std::vector<std::string> expect{
-        sequence_lua_filename,
-        "step_01_action.lua",
-        "step_02_if.lua",
-        "step_03_action.lua",
-        "step_04_elseif.lua",
-        "step_05_action.lua",
-        "step_06_end.lua",
-        "step_07_while.lua",
-        "step_08_action.lua",
-        "step_09_end.lua",
-        "step_10_action.lua"
-    };
-
-    std::vector<std::string> actual = collect_lua_filenames(temp_dir / seq_folder);
-    std::sort(actual.begin(), actual.end());
-
-    REQUIRE(expect.size() == actual.size());
-    REQUIRE(expect == actual);
-}
-
-TEST_CASE("serialize_sequence: loading nonexisting file", "[serialize_sequence]")
-{
-    // empty path
-    REQUIRE_THROWS_AS(load_sequence(""), Error);
-
-    std::filesystem::create_directory(temp_dir);
-
-    // folder 'sequence' does not exist
-    REQUIRE_THROWS_AS(load_sequence(temp_dir / "sequence2"), Error);
-}
-
-TEST_CASE("serialize_sequence: indentation level & type", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    Step step01{Step::type_while};
-    step01.set_label("while condition");
-    Step step02{Step::type_action};
-    step02.set_label("action");
-    Step step03{Step::type_end};
-    step03.set_label("while end");
-
-    Sequence sequence{ "this_is_a_sequence" };
-    sequence.push_back(step01);
-    sequence.push_back(step02);
-    sequence.push_back(step03);
-
-    REQUIRE_NOTHROW(manager.store_sequence(sequence));
-
-    Sequence deserialize_seq = load_sequence(temp_dir / make_sequence_filename(sequence));
-
-    REQUIRE(not deserialize_seq.empty());
-    REQUIRE(deserialize_seq.size() == 3);
-    REQUIRE(deserialize_seq[0].get_indentation_level() == 0);
-    REQUIRE(deserialize_seq[0].get_type() == Step::type_while);
-    REQUIRE(deserialize_seq[1].get_indentation_level() == 1);
-    REQUIRE(deserialize_seq[1].get_type() == Step::type_action);
-    REQUIRE(deserialize_seq[2].get_indentation_level() == 0);
-    REQUIRE(deserialize_seq[2].get_type() == Step::type_end);
-}
-
-TEST_CASE("serialize_sequence: default constructed Step", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    Sequence sequence{ "BlueAsBlood" };
-    sequence.push_back(Step{});
-
-    REQUIRE_NOTHROW(manager.store_sequence(sequence));
-    REQUIRE_NOTHROW(load_sequence(temp_dir / make_sequence_filename(sequence)));
-}
-
-TEST_CASE("serialize_sequence: Sequence label and UID escaping", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    const auto label = "A/\"sequence\"$[<again>]"s;
-    const SequenceName name{ "gabba-gabba_he.Y" };
-
-    Sequence sequence{ label, name };
-    sequence.push_back(Step{});
-
-    const auto seq_folder = make_sequence_filename(sequence);
-
-    if (std::filesystem::exists(temp_dir / seq_folder))
-        std::filesystem::remove_all(temp_dir / seq_folder);
-
-    auto before = collect_filenames(temp_dir);
-
-    REQUIRE_NOTHROW(manager.store_sequence(sequence));
-
-    auto after = collect_filenames(temp_dir);
-
-    std::vector<std::string> new_filenames;
-    std::set_difference(after.begin(), after.end(),
-                        before.begin(), before.end(),
-                        std::back_inserter(new_filenames));
-
-    REQUIRE(new_filenames.size() == 1);
-    REQUIRE(new_filenames[0] == seq_folder);
-
-    Sequence deserialize_seq = load_sequence(temp_dir / new_filenames[0]);
-    REQUIRE(sequence.get_label() == deserialize_seq.get_label());
-    REQUIRE(sequence.get_name() == deserialize_seq.get_name());
-}
-
-TEST_CASE("serialize_sequence: simple step setup", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    Step step_action{Step::type_action};
-    step_action.set_script("a = preface .. 'Bob'");
-
-    Sequence seq{ "seq_with_simple_step_setup" };
-    seq.push_back(step_action);
-    seq.set_step_setup_script("preface = 'Alice calls '");
-
-    manager.store_sequence(seq);
-
-    Sequence seq_deserialized = load_sequence(temp_dir / make_sequence_filename(seq));
-
-    REQUIRE(seq_deserialized.get_step_setup_script() == "preface = 'Alice calls '");
-}
-
-TEST_CASE("serialize_sequence: complex step setup", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    auto step_setup_script =
-R"(
-a = 'Bob'
-        function test(name) return name .. ' with some funky stuff!'     end
-b = test('Alice') \b\b  c = 4)";
-
-    Step step_action{Step::type_action};
-    step_action.set_script("a = preface .. 'Bob'");
-
-    Sequence seq{ "seq_with_complex_step_setup" };
-    seq.push_back(step_action);
-    seq.set_step_setup_script(step_setup_script);
-
-    manager.store_sequence(seq);
-
-    Sequence seq_deserialized = load_sequence(temp_dir / make_sequence_filename(seq));
-
-    REQUIRE(seq_deserialized.get_step_setup_script() == step_setup_script);
-}
-
-TEST_CASE("serialize_sequence: empty sequence", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    std::string uid{ "empty_seq_with_step_setup" };
-    Sequence seq{ uid };
-
-    SECTION("Deserialize empty sequence (part 1)")
-    {
-        manager.store_sequence(seq);
-        REQUIRE_NOTHROW(load_sequence(temp_dir / make_sequence_filename(seq)));
-    }
-
-    SECTION("Deserialize empty sequence (part 2)")
-    {
-        // Remove previously stored sequence
-        if (std::filesystem::exists(temp_dir / make_sequence_filename(seq)))
-            std::filesystem::remove_all(temp_dir / make_sequence_filename(seq));
-
-        manager.store_sequence(seq);
-
-        Sequence seq_deserialized = load_sequence(temp_dir / make_sequence_filename(seq));
-        REQUIRE(seq_deserialized.empty());
-    }
-}
-
-TEST_CASE("serialize_sequence: sequence with step setup script", "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    std::string uid{ "seq_with_step_setup" };
-
-    Step step_action{Step::type_action};
-    step_action.set_script("a = preface .. 'Bob'");
-
-    Sequence seq{ uid };
-    seq.set_step_setup_script("a = 'Bob is alive'");
-    seq.push_back(step_action);
-
-    // remove previously stored sequence
-    if (std::filesystem::exists(temp_dir / make_sequence_filename(seq)))
-        std::filesystem::remove_all(temp_dir / make_sequence_filename(seq));
-
-    manager.store_sequence(seq);
-
-    Sequence seq_deserialized;
-    seq_deserialized.set_step_setup_script("b = 'Should break the test!'");
-
-    load_sequence_parameters(temp_dir / make_sequence_filename(seq), seq_deserialized);
-    REQUIRE(seq_deserialized.get_step_setup_script() == "a = 'Bob is alive'");
-    REQUIRE(seq_deserialized.empty()); // no steps loaded
-}
-
-TEST_CASE("serialize_sequence: sequence maintainers, timeout & nice name",
-    "[serialize_sequence]")
-{
-    SequenceManager manager{ temp_dir };
-
-    Sequence seq{ "Test sequence with maintainers" };
-
-    // remove previously stored sequence
-    if (std::filesystem::exists(temp_dir / make_sequence_filename(seq)))
-        std::filesystem::remove_all(temp_dir / make_sequence_filename(seq));
-
-    seq.set_maintainers("John Doe john.doe@universe.org; Bob Smith boby@milkyway.edu");
-    seq.set_timeout(task::Timeout{1min});
-
-    manager.store_sequence(seq);
-
-    Sequence seq_deserialized;
-
-    std::filesystem::path path{ temp_dir/make_sequence_filename(seq) };
-    load_sequence_parameters(path, seq_deserialized);
-
-    REQUIRE("John Doe john.doe@universe.org; Bob Smith boby@milkyway.edu"
-        == seq_deserialized.get_maintainers());
-    REQUIRE(task::Timeout{1min} == seq_deserialized.get_timeout());
-    REQUIRE("Test sequence with maintainers" == seq_deserialized.get_label());
 }
